requirements.txt
ctransformers
sentence-transformers
pinecone-client
langchain
pypdf
python-dotenv
python-multipart
fastapi
uvicorn